+++ prbt_hardware_support/test/include/prbt_hardware_support/pilz_modbus_server_mock.h
@@ -90,6 +90,8 @@ public:
    */
   void terminate();
 
+  void setTerminateFlag();
+
   /**
    * @brief Allocates needed resources for running the server.
    *
@@ -112,7 +114,16 @@ public:
   void run();
 
 private:
+  /**
+   * @returns True if the server is told to shutdown via Modbus msg,
+   * false otherwise.
+   */
+  bool shutdownSignalReceived();
+
+private:
   const unsigned int holding_register_size_;
+  //! Index of register for server shutdown signal
+  const unsigned int terminate_register_idx_;
 
 private:
 
@@ -137,12 +148,21 @@ private:
   static constexpr uint32_t RESPONSE_TIMEOUT_IN_SEC       {0};
   static constexpr uint32_t RESPONSE_TIMEOUT_IN_USEC      {10000};
 
+  //! Register value which indicates that server has to shutdown.
+  static constexpr uint16_t TERMINATE_SIGNAL {1};
+
 };
+
+inline void PilzModbusServerMock::setTerminateFlag()
+{
+  ROS_INFO_NAMED("ServerMock", "Set terminate to true.");
+  terminate_ = true;
+}
 
 inline void PilzModbusServerMock::terminate()
 {
   ROS_INFO_NAMED("ServerMock", "Terminate called on ServerMock.");
-  terminate_ = true;
+  setTerminateFlag();
 
   if(thread_.joinable())
   {
@@ -150,6 +170,11 @@ inline void PilzModbusServerMock::terminate()
     thread_.join();
     ROS_DEBUG_NAMED("ServerMock", "Waiting for worker Thread of ServerMock joined.");
   }
+}
+
+inline bool PilzModbusServerMock::shutdownSignalReceived()
+{
+  return readHoldingRegister(terminate_register_idx_, 1).back() == TERMINATE_SIGNAL;
 }
 
 }